[semver:patch] make GOOGLE_COMPUTE_{REGION,ZONE} optional
Fixes #38
The changes in #23 make the initialize command fail if one of
GOOGLE_COMPUTE_ZONE or GOOGLE_COMPUTE_REGION is not set. While having
one of these set implicitly may be useful for some people, it is not
required for all commands, and can also usually be specified via --zone,
--region, or --location.

This will also enable _both_ to be set if the env vars exist.

Also, clarify docs:
- Clarify when it's the env var name vs. the value to use
- Clarify wording so it's clear that these are the _default_ values for
  the CLI to use